internal: consider crc32 input unsigned during computation
CRC algorithms are supposed to work with unsigned bytes.

The prototype of ngli_crc32() remains unchanged as it is a string-helper
and not a CRC32 for buffers.

+++ libnodegl/utils.c
@@ -95,7 +95,7 @@ uint32_t ngli_crc32(const char *s)
 {
     uint32_t crc = ~0;
     for (int i = 0; s[i]; i++) {
-        crc ^= s[i];
+        crc ^= (uint8_t)s[i];
         for (int j = 0; j < 8; j++) {
             const uint32_t mask = -(crc & 1);
             crc = (crc >> 1) ^ (mask & 0xEDB88320);
